+++ tests/bootstrap.php
@@ -17,8 +17,6 @@ $application = new Application(
 );
 $application->initialize();
 
-define('NO_PROMPT', true);
-
 Zend_Registry::set(
     'omeka-cli-application',
     $application